fix: Implement bin.Findbin for non-singularity pkg/ usage
If code in `pkg/` uses `bin.Findbin` then prior to this PR it would
attempt to import `buildcfg'.

`buildcfg` is only available as part of a full build of SingularityCE,
not when other code imports `pkg/`.

Add build tag gated variants of the private functions called by
`bin.FindBin` that do not require `buildcfg`, to avoid this issue.
